Actually use uuid's for deck ids

+++ db/migrate/20190121201635_add_added_by_to_decks.rb
@@ -1,5 +1,5 @@
 class AddAddedByToDecks < ActiveRecord::Migration[5.2]
   def change
-    add_column :decks, :added_by, :integer
+    add_column :decks, :added_by, :bigint
   end
 end
